backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
# django-cryptography==1.1  # Not compatible with Django 5.0, removed for now
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
celery==5.3.4
redis==5.0.1
stripe==7.8.0
requests==2.31.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.26.2
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
argon2-cffi==23.1.0

